Pan camera around, FB should ! draw out of bounds

+++ Software/src/games/mage/entity.h
@@ -1,8 +1,8 @@
 #include <cstdint>
 
 struct Point {
-    uint16_t x;
-    uint16_t y;
+    int32_t x;
+    int32_t y;
 };
 struct Entity {
     char name[16];
